(re-)add minimal README.md directing attention to gh-pages.
I have now worked out where to look for the web pages generated by the
content of the gh-pages branch.  This is currently hardcoded in the
readme as http://astrojs.github.com/sampjs/.  If there is some way
to do that using a non-hardcoded reference to the gh-pages site
associated with the project, that would be better.