Adjust broadcaster example to no longer need explicit registration.
The broadcaster.js script and its client example sendlist.html now
no longer have a Register/Unregister button.  Registration is performed
as required when a send is to be performed.  An interval timer keeps
pinging the hub to see if it's alive, and sets the enabledness of
the broadcast buttons accordingly.  This is much more usable (user
doesn't need to understand about register/send/unregister cycle).
Thanks to Markus for opening my eyes to that fact.