Add runWithConnection and onHubAvailability methods to Connector.
The runWithConnection method lets you do on-demand-only registration,
so there's no need for the register/unregister buttons.
onHubAvailability is a thin wrapper for the ping method which gives
you a hook for when the hub starts or stops.  This is not related
to whether the connector is actually registered, but in conjunction
with runWithConnection that is actually what you want.